test: gate the branchy inlined-callee bridge regression in the synth suite
Add pyre/bench/synth/bridge_branchy_callee.py: a branch inside an inlined callee
(gh#343) whose rare arm guard-fails, exercising the carrier-resume sub-walk's
int-arith inline of the branchy continuation. It prints a deterministic sum and
is gated against pypy by `# pyre-check: max-pypy-ratio=8`. The synthetic suite
times pyre-JIT vs pypy (native backends, startup-subtracted): the fix holds at
1.4-1.7x (dynasm) / 3.3-3.6x (cranelift), while a decline that keeps every
crossing interpreted balloons it to ~30x, so the ratio catches the regression
with margin. wasm is exempt (guest has no time.perf_counter()). Drops the old
bridge_branchy_callee_regression.py run_selfcheck.

check.py dynasm + cranelift.
